+++ osu.Game/Rulesets/Edit/Layers/Selection/HandleContainer.cs
@@ -11,77 +11,77 @@ using osu.Framework.Graphics.Primitives;
 namespace osu.Game.Rulesets.Edit.Layers.Selection
 {
     /// <summary>
-    /// A <see cref="CompositeDrawable"/> that has <see cref="Marker"/>s around its border.
+    /// A <see cref="CompositeDrawable"/> that has <see cref="Handle"/>s around its border.
     /// </summary>
-    public class MarkerContainer : CompositeDrawable
+    public class HandleContainer : CompositeDrawable
     {
         /// <summary>
-        /// Invoked when a <see cref="Marker"/> requires the current drag rectangle.
+        /// Invoked when a <see cref="Handle"/> requires the current drag rectangle.
         /// </summary>
         public Func<RectangleF> GetDragRectangle;
 
         /// <summary>
-        /// Invoked when a <see cref="Marker"/> wants to update the drag rectangle.
+        /// Invoked when a <see cref="Handle"/> wants to update the drag rectangle.
         /// </summary>
         public Action<RectangleF> UpdateDragRectangle;
 
         /// <summary>
-        /// Invoked when a <see cref="Marker"/> has finished updates to the drag rectangle.
+        /// Invoked when a <see cref="Handle"/> has finished updates to the drag rectangle.
         /// </summary>
         public Action FinishCapture;
 
-        public MarkerContainer()
+        public HandleContainer()
         {
             InternalChildren = new Drawable[]
             {
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.TopLeft,
                     Origin = Anchor.Centre
                 },
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.TopCentre,
                     Origin = Anchor.Centre
                 },
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.TopRight,
                     Origin = Anchor.Centre
                 },
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.CentreLeft,
                     Origin = Anchor.Centre
                 },
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.BottomLeft,
                     Origin = Anchor.Centre
                 },
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.CentreRight,
                     Origin = Anchor.Centre
                 },
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.BottomRight,
                     Origin = Anchor.Centre
                 },
-                new Marker
+                new Handle
                 {
                     Anchor = Anchor.BottomCentre,
                     Origin = Anchor.Centre
                 },
-                new CentreMarker
+                new OriginHandle
                 {
                     Anchor = Anchor.Centre,
                     Origin = Anchor.Centre
                 }
             };
 
-            InternalChildren.OfType<Marker>().ForEach(m =>
+            InternalChildren.OfType<Handle>().ForEach(m =>
             {
                 m.GetDragRectangle = () => GetDragRectangle();
                 m.UpdateDragRectangle = r => UpdateDragRectangle(r);
