+++ osu.Game/osu.Game.csproj
@@ -307,10 +307,10 @@
     <Compile Include="Overlays\Profile\Sections\Ranks\DrawableTotalScore.cs" />
     <Compile Include="Overlays\Profile\Sections\Ranks\ScoreModsContainer.cs" />
     <Compile Include="Overlays\Settings\SettingsButton.cs" />
-    <Compile Include="Rulesets\Edit\Layers\Selection\CentreMarker.cs" />
+    <Compile Include="Rulesets\Edit\Layers\Selection\OriginHandle.cs" />
     <Compile Include="Rulesets\Edit\Layers\Selection\HitObjectSelectionBox.cs" />
-    <Compile Include="Rulesets\Edit\Layers\Selection\Marker.cs" />
-    <Compile Include="Rulesets\Edit\Layers\Selection\MarkerContainer.cs" />
+    <Compile Include="Rulesets\Edit\Layers\Selection\Handle.cs" />
+    <Compile Include="Rulesets\Edit\Layers\Selection\HandleContainer.cs" />
     <Compile Include="Rulesets\Edit\Layers\Selection\SelectionInfo.cs" />
     <Compile Include="Rulesets\Edit\Layers\Selection\SelectionLayer.cs" />
     <Compile Include="Screens\Edit\Components\BottomBarContainer.cs" />
